fix(ui): improve single stat computation
The method we used to compute a single stat / gauge value previously had
a few issues:

- If the latest value was part of a numeric column but was null/NaN/not
  defined, the single stat computation would fail and a user would see an
  error message "Could not display single stat because your values are
  non-numeric"
- If there were multiple latest values, an arbitrary selection would be
  made

This commit updates the single stat computation to find the latest
*defined* numeric values.

If multiple latest valid numeric values are found, we will either:

- Display an error message if using the compuation within a single stat
  visualization
- Display nothing if using the computation within a within a line +
  single stat visualization (i.e. display the line vis only)

The behavior is similar if no latest numeric value is found.

This commit also updates the single stat computation to use the
@influxdata/vis `Table` format as an intermediate/parsed representation
of a Flux CSV response. This unlocks the possibility for performance
gains in our CSV parsing. See #13852.

Closes #13824